+++ app/invitations/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import InvitationCenter from "@/components/invitations/InvitationCenter";
+
+export default function InvitationsPage() {
+  return <InvitationCenter />;
+}